+++ source/Public/Disable-SqlDscAudit.ps1
@@ -70,7 +70,7 @@ function Disable-SqlDscAudit
 
     process
     {
-        if ($Force.IsPresent)
+        if ($Force.IsPresent -and -not $Confirm)
         {
             $ConfirmPreference = 'None'
         }